requirements.txt
Flask>=2.1.1
requests>=2.27.1
beautifulsoup4>=4.11.1
python-dotenv>=1.0.0
google-generativeai>=0.3.1
firecrawl-py>=0.0.14 # Add firecrawl library